Authorization done

+++ todo/main.py
@@ -1,23 +1,18 @@
 from flask import Flask, request
-from flask_restplus import Api, Resource, fields, reqparse
+from flask_restplus import Api, Resource, fields, reqparse, marshal
 import pymysql
 import datetime
+from base64 import b64decode, b64encode
 
 app = Flask(__name__)
 
-# authorizations={
-#      'oauth2': {
-#         'type': 'oauth2',
-#         'in': 'header',
-#         'flow': 'password',
-#          'tokenUrl': '/authToken',
-#         'authorizationUrl': 'https://somewhere.com/auth',
-#         'scopes': {
-#             'read': 'Grant read-only access',
-#             'write': 'Grant read-write access',
-#         }
-#      }
-# }
+authorizations={
+     'basic': {
+        'type': 'basic',
+         'in': 'header',
+          'name': 'Authorization'
+     }
+}
 #Connect to the database
 connection = pymysql.connect(host='localhost',
                              user='krishna',
@@ -36,13 +31,22 @@ tasks table
 | due_date | date          | YES  |     | NULL        |                |        
 | status   | varchar(20)   | NO   |     | Not Started |                |      
 +----------+---------------+------+-----+-------------+----------------+
+
+user table
++----------+--------------+------+-----+---------+-------+
+| Field    | Type         | Null | Key | Default | Extra |
++----------+--------------+------+-----+---------+-------+
+| username | varchar(200) | NO   | PRI | NULL    |       |
+| password | varchar(200) | NO   |     | NULL    |       |
+| role     | varchar(20)  | YES  |     | read    |       |
++----------+--------------+------+-----+---------+-------+
 '''
 
 api = Api(version='1.0', title='TodoMVC API',
     description='A simple TodoMVC API',
-    # authorizations=authorizations
+    security='basic',
+    authorizations=authorizations,
 )
-# api.add_namespace(ns1)                 #Adding user namespace to the api
 
 ns = api.namespace('todos', description='TODO operations')
 todo = api.model('Todo', {
@@ -55,6 +59,11 @@ api.add_namespace(ns)
 
 api.init_app(app)
 
+invalid=api.model('Failed',{                    #Model for marshal decorator indicating invalid credentials
+    'message':fields.String(default='Invalid Credentials')
+})
+
+
 
 #Check if the string date_text is a valid date in iso format
 def isDateValid(date_text):
@@ -63,6 +72,37 @@ def isDateValid(date_text):
         return True
     except :
         return False
+
+
+#Helper class for user operations
+class UserDAO(object):
+    def __init__(self):
+        self.user={}
+        pass
+
+    def login(self,user):
+
+        cursor=connection.cursor()
+        sql="select * from user where username=%s"
+        cursor.execute(sql,(user['username']))
+        dbUser=cursor.fetchone()
+        cursor.close()
+        if(dbUser==None or dbUser['password']!=user['password']):
+            return 0
+        self.user=dbUser
+        return 1
+
+    def isAllowed(self,user,role):                              #Check user's role against the caller method's required access
+        if(user==None):
+            return {"message":"Invalid Credentials"}, 401
+        if(self.login(user)==0):                                #If user login is not success return 0
+            return {"message":"Invalid Credentials"}, 401
+
+        if(self.user['role']=='read' and role=='write'):        #If login success but user doesn't have access
+            return {"message":"Access Restricted"},403
+
+        return 1,200                        #When login is success and allowed to access method
+            
 
 #Helper class for operations on todo list
 class TodoDAO(object):
@@ -151,23 +191,34 @@ class TodoDAO(object):
 
 
 DAO = TodoDAO()
+userDAO=UserDAO()
+
 
 @ns.route('/')
+@ns.response(code=401, model=invalid, description='Invalid Credentials')
 class TodoList(Resource):
     '''Shows a list of all todos, and lets you POST to add new tasks'''
     @ns.doc('list_todos')
-    @ns.marshal_list_with(todo)
+    @ns.response(code=200, model=todo, description='Success')
     def get(self):
         '''List all tasks'''
-        # print(current_user )
-        return DAO.getAll()
+        status=userDAO.isAllowed(request.authorization,'read')
+        if(status[1]!=200):
+            return status
+
+        return marshal(DAO.getAll(), todo), 200
+
 
     @ns.doc('create_todo')
     @ns.expect(todo)            #Post request expects a todo JSON object
-    @ns.marshal_with(todo, code=201)
+    @ns.response(code=201, model=todo, description='Success')
+    @ns.response(code=403, model=invalid, description='Access Restricted')
     def post(self): 
         '''Create a new task'''
         data=api.payload 
+        status=userDAO.isAllowed(request.authorization,'write')
+        if(status[1]!=200):
+            return status
         
         if(data['status'] not in ['Finished','Not Started','In Progress']):    #If status doesn't match expected value   
             return {"message":"Invalid status value provided"}, 422
@@ -175,7 +226,7 @@ class TodoList(Resource):
         if(not isDateValid(data['due_date'])):           #If date provided is not a valid date
             return {"message":"Invalid date value provided"}, 422
 
-        return DAO.create(api.payload), 201
+        return marshal(DAO.create(api.payload),todo), 201
 
 #Parser for due date 
 dateParser=reqparse.RequestParser()
@@ -187,12 +238,17 @@ class TodoListDue(Resource):
     
     @ns.doc('list_todos by due date')
     @ns.expect(dateParser)
-    @ns.marshal_list_with(todo)
+    @ns.response(code=200, model=todo, description='Success')
+    @ns.response(code=401, model=invalid, description='Invalid Credentials')
     def get(self):
         '''List all tasks by due_date'''
+        status=userDAO.isAllowed(request.authorization,'read')      #Check if user is authorized to access
+        if(status[1]!=200):
+            return status
+
         args = dateParser.parse_args()
         date=args['due_date']
-        return DAO.getAll(date), 200
+        return marshal(DAO.getAll(date),todo), 200
 
 
 #Route and resource for fetching tasks from list which are overdue
@@ -200,10 +256,15 @@ class TodoListDue(Resource):
 class TodoListOverdue(Resource):
 
     @ns.doc('list_todos which are overdue')
-    @ns.marshal_list_with(todo)
+    @ns.response(code=200, model=todo, description='Success')
+    @ns.response(code=401, model=invalid, description='Invalid Credentials')
     def get(self):
         '''List all tasks that are overdue'''
-        return DAO.getOverdue(), 200
+        status=userDAO.isAllowed(request.authorization,'read')      #Check if user is authorized to access
+        if(status[1]!=200):
+            return status
+
+        return marshal(DAO.getOverdue(),todo), 200
 
 
 #Route and resource for fetching tasks from list which are finished.
@@ -211,35 +272,54 @@ class TodoListOverdue(Resource):
 class TodoListOverdue(Resource):
 
     @ns.doc('list_todos which are finished')
-    @ns.marshal_list_with(todo)
+    @ns.response(code=200, model=todo, description='Success')
+    @ns.response(code=401, model=invalid, description='Invalid Credentials')
     def get(self):
         '''List all tasks that are finished'''
-        return DAO.getFinished(), 200
+        status=userDAO.isAllowed(request.authorization,'read')      #Check if user is authorized to access
+        if(status[1]!=200):
+            return status
+
+        return marshal(DAO.getFinished(),todo), 200
 
 #Routes for single tasks
 @ns.route('/<int:id>')
 @ns.response(404, 'Todo not found')
+@ns.response(code=200, model=todo, description='Success')
+@ns.response(code=401, model=invalid, description='Invalid Credentials')
+@ns.response(code=403, model=invalid, description='Access Restricted')
 @ns.param('id', 'The identifier')
 class Todo(Resource):
     '''Show a single todo item and lets you delete them'''
     @ns.doc('get_todo')
-    @ns.marshal_with(todo)      #Filters the output based on the fields provided in the todo model
+ 
     def get(self, id):
         '''Fetch a given resource'''
-        return DAO.get(id)
+        status=userDAO.isAllowed(request.authorization,'read')      #Check if user is authorized to access
+        if(status[1]!=200):
+            return status
+
+        return marshal(DAO.get(id),todo)
 
     @ns.doc('delete_todo')
     @ns.response(204, 'Todo deleted')
     def delete(self, id):
         '''Delete a task given its identifier'''
+        status=userDAO.isAllowed(request.authorization,'write')      #Check if user is authorized to access
+        if(status[1]!=200):
+            return status
+
         DAO.delete(id)
         return '', 204
 
     @ns.expect(todo)
-    @ns.marshal_with(todo)
     def put(self, id):
         '''Update a task given its identifier'''
-        return DAO.update(id, api.payload)
+        status=userDAO.isAllowed(request.authorization,'write')      #Check if user is authorized to access
+        if(status[1]!=200):
+            return status
+
+        return marshal(DAO.update(id, api.payload),todo), 200
 
 
 #Parser for TodoStatus put request formData
@@ -248,28 +328,27 @@ parser.add_argument('status', type=str, help='Status of the task', location='for
 
 #Resource to update a task's status
 @ns.route('/<int:id>/status', methods=['PUT'])
-@ns.response(200, 'Success')
+@ns.response(code=200, model=todo, description='Success')
+@ns.response(code=401, model=invalid, description='Invalid Credentials')
 @ns.response(404, 'Todo not found')
 @ns.response(422, 'Incorrect value provided')
+@ns.response(code=403, model=invalid, description='Access Restricted')
 @ns.param('id', 'The identifier')
 class TodoStatus(Resource):
 
     @ns.expect(parser)
-    @ns.marshal_with(todo)
     def put(self, id):
         '''Update the status of the given task. Can take values- 'Not Started','Finished','In progress' '''
+        status=userDAO.isAllowed(request.authorization,'write')      #Check if user is authorized to access
+        if(status[1]!=200):
+            return status
+
         args = parser.parse_args()
         status=args['status']
         if(status not in ['Finished','Not Started','In Progress']):
             return {"message":"Invalid value provided"}, 422
 
-        return DAO.update(id,args['status'],True)
-
-
-# @api.route('/authToken')
-# class Auth(Resource):
-#     def get(self):
-#         pass
+        return marshal(DAO.update(id,args['status'],True),todo), 200
 
 if __name__== "__main__":
     app.run(debug=True)
